fix: /model에서 타입 정의 통일화
